Allocates within a single page group.
The code to allocate new pages when we run out of space in the current
page group has not been written yet.  So when we run out of memory in that
page, it simply returns NULL and gives you the middle finger on the
terminal output.

The 'maxfree' variable was removed from the page group.  To simplify the
code, chunks will not be allocated that reside between chunks that are
still in use.  Once a chunk is marked as free, if that chunk is at the top
of the page group's allocation stack, all freed chunks will be released.
So all we really care about is "how many bytes can be allocated from this
page group" by only looking at the distance between 'top' and the end of
the page group.

Debug messages were added and existing debug messages were clarified and
cleaned up.